\name{daylength_hours}
\alias{daylength_hours}
\title{Daylength}
\description{Astronomical daylength (hours) from latitude and day of year (Allen et al., 1998, FAO-56).}
\usage{daylength_hours(lat_deg, doy)}
\arguments{\item{lat_deg}{Latitude (decimal degrees, +N).}
\item{doy}{Day of year (1-366).}
}
\value{Daylength (h).}
\examples{daylength_hours(35, 180)}
