\name{spacsysr-package}
\alias{spacsysr-package}
\title{spacsysr: Core SPACSYS Biogeochemistry in R}
\description{R implementations of the core soil biogeochemical process modules of the SPACSYS model (Wu, 2022, ver. 6.00): response functions, decomposition, nitrification/denitrification, N2O/NO/NH3/CH4 emissions, plant respiration, soil water and a daily driver.}
\details{Start with \code{\link{spacsys_run}} and the N2O vignette. This package implements the well-documented soil biogeochemistry core of SPACSYS; it is not a full port of the farm-scale model (see README).}
\references{Wu, L. (2022). SPACSYS - FARMSYS: Soil-Plant-Atmosphere Continuum System Model (ver. 6.00) - Technical Manual. Rothamsted Research, North Wyke, UK.}
